+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 Key: [+] added, [-] removed, [~] modified, [!] important.
 
+## 2019_06-11
++ [+] Added [__advice__] command using api.adviceslip.com which is quick and easy.
+
 ## 2019-06-10
 + [~] Updated dependencies to newest possible without breaking other dependencies @JDA && OkHttp
 + [~] Refactored ApplicationProgrammingInterface to simply Api and put it into its own package along with an ApiManager class.